Add XML benchmarks
